backend/requirements.txt
# CoSauce Platform Backend Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
httpx>=0.26.0
openai>=1.0.0
python-docx>=1.1.0
python-pptx>=0.6.21
aiosqlite>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
